pyproject.toml
[tool.poetry]
name = "python-graph-viewer"
version = "0.1.0"
description = ""
authors = ["hakohumi <mihumihu32@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.22.4"
matplotlib = "^3.5.2"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
